pyproject.toml
[project]
name = "legal-assist-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.11.0",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-huggingface>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.2",
    "pymilvus>=2.6.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
]
